:+1: init_envの関数分け
