+++ header/environ.h
@@ -3,6 +3,9 @@
 
 int add_env(t_condition *condition, char *env_str);
 t_item *touch_t_item();
+t_item *set_tail(t_item *head);
+t_item *serch_dup_key(t_condition *condition, char *key);
+
 
 
 # endif
